+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Artificial emotional intelligence".
 - Tests for "Park pilot".
 - Tests for "Buzzle".
+- Tests for "How time flies".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".